Implemented term functions in GO.ontology.GeneOntologyNX.
Largely, implemented the logic for adding, storing, and removing terms from ontology, as well as checking for consistent storage states. For the last item, I added a new exception, InternalStorageInconsistentError, to be raised if an ontology discovers its internal storage structures report differently whether a term is or is not stored.

I also added goid as an attribute of GOTerm. This serves as a reference to the identifier attribute, but is more understandable in the context of the gene ontology.